+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.0]
+### Added
+- Macro support. Any macro can be specified in the form of `{MACRO:ARGUMENT}` inside your Markdown files.
+  Check out the updated README for more info.
+
+
 ## [0.1.2] - 2017-08-21
 ### Changed
 - Updated documentation.